Begin unifying loading of resource files

+++ ui/renderer/text_renderer.h
@@ -29,7 +29,7 @@ public:
     TextRenderer() = default;
     void setup(float width, float height, std::string main_font_name, int font_size);
     void renderText(Buffer& buffer, float scroll_x, float scroll_y);
-    void resize(int new_width, int new_height);
+    void resize(float new_width, float new_height);
     void setCursorPositions(Buffer& buffer, float cursor_x, float cursor_y, float drag_x,
                             float drag_y);
     void parseBuffer(Buffer& buffer);
